Modifications to how History counters are saved
Only track History for Quiet Moves now
Now clearing History between games, not between searches
Before each search, cut the History in fourths
Cleaned up some logic for Futility Pruning
Cleaned up some logic for extending depth 0 if in check
Moved Old Versions of Ethereal to https://andygrant.github.io/

QuietHistory2 @ 5.0+0.05 with 128MB Hash
Base                Alpha   Beta    Games   Test Elo  Base Elo  Status
TTCapture+Bound3    0.01    0.01    13350   2728.38   2713.91   PASSED

MATCHUPS FOR CANDIDATE
Opponent        Wins    Draws   Losses  Games   Score   ELO      +      -
Floyd0.9        785     125     190     1100    77.05%  2800.35  25.42  23.55
Laser1.0        199     134     267     600     44.33%  2687.45  27.79  28.30
Sloppy0.2.2     323     103     124     550     68.09%  2748.67  32.28  30.24
WyldChess1.4    330     115     105     550     70.45%  2760.97  33.16  30.75
DiscoCheck4.03  152     124     274     550     38.91%  2633.63  29.29  30.43
RedQueen1.1.98  232     117     201     550     52.82%  2691.60  29.29  29.01
Ethereal8.15    238     203     109     550     61.73%  2697.04  30.56  29.35
Tucano3.0       353     116     81      550     74.73%  2807.33  35.20  31.98

MATCHUPS FOR BASE
Opponent        Wins    Draws   Losses  Games   Score   ELO      +      -
Floyd0.9        714     112     224     1050    73.33%  2765.73  24.59  23.06
Laser1.0        533     373     694     1600    44.97%  2691.92  17.04  17.21
Sloppy0.2.2     458     150     192     800     66.62%  2737.09  26.22  24.96
WyldChess1.4    664     203     233     1100    69.59%  2753.82  22.93  21.80
DiscoCheck4.03  386     318     796     1500    36.33%  2614.56  18.04  18.56
RedQueen1.1.98  311     137     302     750     50.60%  2676.17  24.93  24.89
Ethereal8.15    312     265     173     750     59.27%  2679.14  25.70  25.01
Tucano3.0       506     158     136     800     73.12%  2792.89  28.24  26.26

QuietHistory2 @ 30.0+0.3 with 128MB Hash
Base                Alpha   Beta    Games   Test Elo  Base Elo  Status
TTCapture+Bound3    0.01    0.10    15550   2715.42   2703.38   PASSED

MATCHUPS FOR CANDIDATE
Opponent        Wins    Draws   Losses  Games   Score   ELO      +      -
Floyd0.9        403     115     182     700     65.79%  2703.57  27.87  26.52
Laser1.0        267     212     321     800     46.62%  2703.51  24.06  24.28
Sloppy0.2.2     403     152     145     700     68.43%  2751.38  28.59  26.95
WyldChess1.4    645     289     216     1150    68.65%  2746.18  22.19  21.18
DiscoCheck4.03  228     197     325     750     43.53%  2666.81  24.89  25.36
RedQueen1.1.98  466     328     606     1400    45.00%  2637.14  18.21  18.41
Ethereal8.15    320     270     110     700     65.00%  2721.54  27.68  26.42
Tucano3.0       909     377     214     1500    73.17%  2793.26  20.40  19.35

MATCHUPS FOR BASE
Opponent        Wins    Draws   Losses  Games   Score   ELO      +      -
Floyd0.9        398     106     196     700     64.43%  2693.19  27.55  26.35
Laser1.0        261     256     383     900     43.22%  2679.61  22.74  23.15
Sloppy0.2.2     387     170     143     700     67.43%  2743.40  28.30  26.78
WyldChess1.4    605     297     248     1150    65.52%  2721.54  21.56  20.76
DiscoCheck4.03  191     197     312     700     41.36%  2651.33  25.84  26.53
RedQueen1.1.98  487     381     632     1500    45.17%  2638.31  17.59  17.77
Ethereal8.15    301     281     118     700     63.07%  2706.99  27.27  26.19
Tucano3.0       912     369     219     1500    73.10%  2792.67  20.39  19.34

+++ src/types.h
@@ -127,7 +127,7 @@ typedef struct PVariation {
     
 } PVariation;
 
-typedef uint16_t HistoryTable[COLOUR_NB][SQUARE_NB][SQUARE_NB][2]; 
+typedef int HistoryTable[COLOUR_NB][SQUARE_NB][SQUARE_NB][2]; 
 
 typedef struct MovePicker {
     int isQuiescencePick, stage, split;
